+++ src/pyroads/homogeneous_segmentation/__init__.py
@@ -1,2 +1,10 @@
 from ._seg_shs import segment_ids_to_maximize_spatial_heterogeneity as segment_ids_to_maximize_spatial_heterogeneity
 from ._seg_mcv import segment_ids_to_minimize_coefficient_of_variation as segment_ids_to_minimize_coefficient_of_variation
+from .._dataframe import supports_pandas_and_polars
+
+segment_ids_to_maximize_spatial_heterogeneity = supports_pandas_and_polars(
+	segment_ids_to_maximize_spatial_heterogeneity
+)
+segment_ids_to_minimize_coefficient_of_variation = supports_pandas_and_polars(
+	segment_ids_to_minimize_coefficient_of_variation
+)